Readme.md
This code completes the assignment for module 3. Basic script that reads the CSV file in the Resource folder and prints the requested information. in addition it also exports the results into a .txt file that is located in the analysis folder. Pybank and Pypoll scripts were assisted by sub Bill. Gave me some key examples that guided me to the solution for both scripts.